Added feature UNFOCUS and refactored
